refactor : configuration 패키지 정리
