+++ src/main/java/com/yapp/artie/global/config/S3Config.java
@@ -1,4 +1,4 @@
-package com.yapp.artie.global.config.aws;
+package com.yapp.artie.global.config;
 
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
